Split assertion module
